refactor(sync): extract release_activation + release state machine (py)
P2.B.1 (Python half). Move the release-archive activation subsystem
(cross-process lock + generation tree + staging + extract-meta + zip
validate/extract + prune) to sync/release_activation, and the release
state machine (CacheMeta + download + manifest verify + sync_release) to
sync/release. Shared types (RepoSpec/ReleaseArchiveSpec/SyncResult) move
to sync/_types (back-edge avoidance). data/sync.py becomes a re-export
barrel for these and retains the GameData-pair state machine (moves to
sync/gamedata_pair in P2.B.2).

Repoint the test_sync_release patches whose callers relocated, following
the caller-timing principle: check_latest_release/download_release_asset/
_get_cascading -> sync.release; _safe_extract_zip/_ACTIVATION_LOCK_* ->
sync.release_activation. The sync_release / sync_release_archive patches
are intentionally NOT repointed here -- their caller (sync_release_archive)
still lives in data/sync, so the barrel keeps them working until P2.B.2.

No behavior change; 369 passed / 23 skipped.
